Rename uiActionsCreators to uiActionCreators
